backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
supabase>=2.0.0

# LangChain / LangGraph
langchain>=0.2.0
langchain-openai>=0.1.0
langgraph>=0.2.0

# Document extraction
pdfplumber>=0.11.0
pytesseract>=0.3.10
Pillow>=10.0.0

# HTTP + utilities
httpx>=0.27.0
python-multipart>=0.0.9
